+++ CHANGES.md
@@ -8,6 +8,8 @@
 
 - Added analysis.plotSpikeStats() func to plot stats of cell rates, ISI CVs and synchronies 
 
+- Added cfg.compactConnFormat option to replace conns dict format with compact list format 
+
 - Added option to plotConn() and plot2Dnet() to load data from compact format json files 
 
 - Adapted python2 code so conversion to python3 via 2to3 command works straight away